+++ LeProjet/LeProjet/files/src/Player.cpp
@@ -31,8 +31,8 @@ int Player::UpdateUnit(EnvItem* envItems, size_t envItemsLength, float delta)
         hitAnySide = 0;
 
         // Optimisation to check collision only if the player is near the environment item
-        if (fabs(p->x - ei->rect.x) > 200
-            && fabs(p->y - ei->rect.y) > 200)
+        if (fabs((double) (p->x - ei->rect.x)) > 200
+            && fabs((double) (p->y - ei->rect.y)) > 200)
         {
             continue;
         }
